+++ b4n/ui/views/shell/mod.rs
@@ -1,4 +1,5 @@
 pub use self::view::*;
 
 mod bridge;
+mod keys;
 mod view;